desktop/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.33.0
psutil==6.1.1
py-cpuinfo==9.0.0
GPUtil==1.4.0
pywin32==308
WMI==1.5.1
pywebview==6.1
pyinstaller==6.11.1
